chore: consolidate util modules to one module
