+++ src/autotarcompress/decrypt_manager.py
@@ -15,7 +15,7 @@ from cryptography.hazmat.primitives.ciphers.aead import AESGCM
 
 from autotarcompress.base_manager import BaseCryptoManager
 from autotarcompress.metadata import get_file_hash, update_decrypted_hash
-from autotarcompress.utils.hash_utils import calculate_sha256
+from autotarcompress.utils import calculate_sha256
 
 
 class DecryptManager(BaseCryptoManager):
