--- src/autotarcompress/utils/__init__.py
@@ -1,13 +1,0 @@
-"""Utility package for backup operations."""
-
-from .utils import (
-    ensure_backup_folder,
-    is_pv_available,
-    validate_and_expand_paths,
-)
-
-__all__ = [
-    "ensure_backup_folder",
-    "is_pv_available",
-    "validate_and_expand_paths",
-]
